analytics/requirements.txt
# StockMind — Microservicio de Analítica Predictiva
# Dependencias Python

# Framework web
flask==3.0.0
flask-cors==4.0.0

# Conexión a MySQL
mysql-connector-python==8.2.0

# Análisis de datos y predicción
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.4.0

# Variables de entorno
python-dotenv==1.0.0

# Servidor WSGI para producción (opcional)
gunicorn==21.2.0
